Remove johnson's algorithm

+++ graph/source/apsp/johnson.h
@@ -68,9 +68,9 @@ struct graph_cuda_t {
 
 // If CUDA_FOUND is false the function won't be defined. When using this function make sure it is not
 // used when CUDA_FOUND is false or linker errors will occur. Use constexpr if or conditional preprocessor branches.
-void johnson_cuda_impl(graph_cuda_t<std::vector<int>, std::vector<edge_t>>& gr, std::vector<double>& output);
+void johnson_cuda_impl(graph_cuda_t<std::vector<int>, std::vector<edge_t>>& gr, std::vector<float>& output);
 
-void johnson_parallel_impl(graph_t& gr, std::vector<double>& output);
+void johnson_parallel_impl(graph_t& gr, std::vector<float>& output);
 } // namespace apsp
 
 #endif
